Change Result::then for Result::call
The name `then` suggests that the query will be processed in the future,
and that is not accurate, since it's not a promise.
The `name` call serves better to indicate that the command has already
been handled and we have the result, and that it justs executes a
callback passing along the unwrapped result.